Fix and simplify sync of key info to columns table
The previous CTE approach modified only columns that were recorded in
the statistics table, which was a bug for some DROP statements, when
statistic data could already have been removed and as a result the column
data was not correctly updated.

The approach with a subquery solves that by updating all column records
for a given table (resetting "column_key" to "" when there is no column key).